Support for AlphaFold and user-defined PDB files

+++ AsyncAPI/Consumers/FASTAResultConsumer.cs
@@ -16,6 +16,7 @@ public class FASTAResultConsumer : IConsumer<FASTAResult>
 {
     private readonly ILogger<FASTAResultConsumer> _logger;
     private readonly ISubmissionRepository _submissionRepository;
+    private readonly IAlphaFoldReceptorRepository _alphaFoldReceptorRepository;
     private readonly IConnectionMultiplexer _redis;
     private readonly IConfiguration _configuration;
     private readonly IDockingPrepService _dockingPrepService;
@@ -23,12 +24,14 @@ public class FASTAResultConsumer : IConsumer<FASTAResult>
 
     public FASTAResultConsumer(ILogger<FASTAResultConsumer> logger,
                                ISubmissionRepository submissionRepository,
+                               IAlphaFoldReceptorRepository alphaFoldReceptorRepository,
                                IConnectionMultiplexer redis, IConfiguration configuration,
                                IDockingPrepService dockingPrepService,
                                IFileService fileService)
     {
         _logger = logger;
         _submissionRepository = submissionRepository;
+        _alphaFoldReceptorRepository = alphaFoldReceptorRepository;
         _redis = redis;
         _configuration = configuration;
         _dockingPrepService = dockingPrepService;
@@ -54,33 +57,47 @@ public class FASTAResultConsumer : IConsumer<FASTAResult>
             return;
         }
 
-        var submission = await _submissionRepository.GetAsync(taskInfo.submissionId);
-        if (submission is null)
+        if (taskInfo.type == FASTATaskType.UserPDB) 
         {
-            _logger.LogError($"Could not find submission of FASTATask. Submission: {taskInfo.submissionId}");
-            return;
-        }
+            var submission = await _submissionRepository.GetAsync(taskInfo.submissionId);
+            if (submission is null)
+            {
+                _logger.LogError($"Could not find submission of FASTATask. Submission: {taskInfo.submissionId}");
+                return;
+            }
 
-        var receptor = submission.receptors.FirstOrDefault(r => r.guid == taskInfo.receptorGuid);
-        if (receptor is null)
+            var receptor = submission.receptors.FirstOrDefault(r => r.guid == taskInfo.receptorGuid);
+            if (receptor is null)
+            {
+                _logger.LogError($"Could not find receptor of FASTATask. Receptor: {taskInfo.receptorGuid}");
+                return;
+            }
+
+            receptor.FASTA = model.FASTA;
+            var maxSize = int.Parse(_configuration.GetSection("Limitations")["MaxReceptorSize"]);
+
+            if (model.FASTA.Length > maxSize)
+            {
+                receptor.status = ReceptorFileStatus.TooBig;
+                receptor.affinity = -1;
+                _fileService.RemoveFile(receptor.file!);
+                receptor.file = null;
+                await _submissionRepository.UpdateReceptor(submission, receptor);
+                return;
+            }
+
+            await _dockingPrepService.PrepareForDocking(submission, receptor);
+        } else 
         {
-            _logger.LogError($"Could not find receptor of FASTATask. Receptor: {taskInfo.receptorGuid}");
-            return;
+            var receptor = await _alphaFoldReceptorRepository.GetByUnitProtID(taskInfo.UnitProtID);
+            if (receptor is null)
+            {
+                _logger.LogError($"Could not find AlphaFold receptor of FASTATask. Receptor: {taskInfo.UnitProtID}");
+                return;
+            }
+
+            receptor.FASTA = model.FASTA;
+            await _alphaFoldReceptorRepository.UpdateAsync(receptor.id!, receptor);
         }
-
-        receptor.FASTA = model.FASTA;
-        var maxSize = int.Parse(_configuration.GetSection("Limitations")["MaxReceptorSize"]);
-
-        if (model.FASTA.Length > maxSize)
-        {
-            receptor.status = ReceptorFileStatus.TooBig;
-            receptor.affinity = -1;
-            _fileService.RemoveFile(receptor.file!);
-            receptor.file = null;
-            await _submissionRepository.UpdateReceptor(submission, receptor);
-            return;
-        }
-
-        await _dockingPrepService.PrepareForDocking(submission, receptor);
     }
 }
